// server/generated/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  CountryCode: { input: any; output: any; }
  FuzzyDateInt: { input: any; output: any; }
  Json: { input: any; output: any; }
};
// ...
export type MediaList = {
  /** Map of advanced scores with name keys */
  advancedScores: Maybe<Scalars['Json']['output']>;
  /** When the entry was completed by the user */
  completedAt: Maybe<FuzzyDate>;
  /** When the entry data was created */
  createdAt: Maybe<Scalars['Int']['output']>;
  /** Map of booleans for which custom lists the entry are in */
  customLists: Maybe<Scalars['Json']['output']>;
  /** If the entry shown be hidden from non-custom lists */
  hiddenFromStatusLists: Maybe<Scalars['Boolean']['output']>;
  /** The id of the list entry */
  id: Scalars['Int']['output'];
  media: Maybe<Media>;
  /** The id of the media */
  mediaId: Scalars['Int']['output'];
  /** Text notes */
  notes: Maybe<Scalars['String']['output']>;
  /** Priority of planning */
  priority: Maybe<Scalars['Int']['output']>;
  /** If the entry should only be visible to authenticated user */
  private: Maybe<Scalars['Boolean']['output']>;
  /** The amount of episodes/chapters consumed by the user */
  progress: Maybe<Scalars['Int']['output']>;
  /** The amount of volumes read by the user */
  progressVolumes: Maybe<Scalars['Int']['output']>;
  /** The amount of times the user has rewatched/read the media */
  repeat: Maybe<Scalars['Int']['output']>;
  /** The score of the entry */
  score: Maybe<Scalars['Float']['output']>;
  /** When the entry was started by the user */
  startedAt: Maybe<FuzzyDate>;
  /** The watching/reading status */
  status: Maybe<MediaListStatus>;
  /** When the entry data was last updated */
  updatedAt: Maybe<Scalars['Int']['output']>;
  user: Maybe<User>;
  /** The id of the user owner of the list entry */
  userId: Scalars['Int']['output'];
};


/** List of anime or manga */
export type MediaListCustomListsArgs = {
  asArray: InputMaybe<Scalars['Boolean']['input']>;
};


/** List of anime or manga */
export type MediaListScoreArgs = {
  format: InputMaybe<ScoreFormat>;
};

/** List of anime or manga */
export type MediaListCollection = {
  /**
   * A map of media list entry arrays grouped by custom lists
   * @deprecated Not GraphQL spec compliant, use lists field instead.
   */
  customLists: Maybe<Array<Maybe<Array<Maybe<MediaList>>>>>;
  /** If there is another chunk */
  hasNextChunk: Maybe<Scalars['Boolean']['output']>;
  /** Grouped media list entries */
  lists: Maybe<Array<Maybe<MediaListGroup>>>;
  /**
   * A map of media list entry arrays grouped by status
   * @deprecated Not GraphQL spec compliant, use lists field instead.
   */
  statusLists: Maybe<Array<Maybe<Array<Maybe<MediaList>>>>>;
  /** The owner of the list */
  user: Maybe<User>;
};


/** List of anime or manga */
export type MediaListCollectionCustomListsArgs = {
  asArray: InputMaybe<Scalars['Boolean']['input']>;
};


/** List of anime or manga */
export type MediaListCollectionStatusListsArgs = {
  asArray: InputMaybe<Scalars['Boolean']['input']>;
};

/** List group of anime or manga entries */
export type MediaListGroup = {
  /** Media list entries */
  entries: Maybe<Array<Maybe<MediaList>>>;
  isCustomList: Maybe<Scalars['Boolean']['output']>;
  isSplitCompletedList: Maybe<Scalars['Boolean']['output']>;
  name: Maybe<Scalars['String']['output']>;
  status: Maybe<MediaListStatus>;
};
// ...
export enum MediaListStatus {
  /** Finished watching/reading */
  Completed = 'COMPLETED',
  /** Currently watching/reading */
  Current = 'CURRENT',
  /** Stopped watching/reading before completing */
  Dropped = 'DROPPED',
  /** Paused watching/reading */
  Paused = 'PAUSED',
  /** Planning to watch/read */
  Planning = 'PLANNING',
  /** Re-watching/reading */
  Repeating = 'REPEATING'
}
// ...
export enum MediaStatus {
  /** Ended before the work could be finished */
  Cancelled = 'CANCELLED',
  /** Has completed and is no longer being released */
  Finished = 'FINISHED',
  /** Version 2 only. Is currently paused from releasing and will resume at a later date */
  Hiatus = 'HIATUS',
  /** To be released at a later date */
  NotYetReleased = 'NOT_YET_RELEASED',
  /** Currently releasing */
  Releasing = 'RELEASING'
}
// ...
export type User = {
  /** The bio written by user (Markdown) */
  about: Maybe<Scalars['String']['output']>;
  /** The user's avatar images */
  avatar: Maybe<UserAvatar>;
  /** The user's banner images */
  bannerImage: Maybe<Scalars['String']['output']>;
  bans: Maybe<Scalars['Json']['output']>;
  /** When the user's account was created. (Does not exist for accounts created before 2020) */
  createdAt: Maybe<Scalars['Int']['output']>;
  /** Custom donation badge text */
  donatorBadge: Maybe<Scalars['String']['output']>;
  /** The donation tier of the user */
  donatorTier: Maybe<Scalars['Int']['output']>;
  /** The users favourites */
  favourites: Maybe<Favourites>;
  /** The id of the user */
  id: Scalars['Int']['output'];
  /** If the user is blocked by the authenticated user */
  isBlocked: Maybe<Scalars['Boolean']['output']>;
  /** If this user if following the authenticated user */
  isFollower: Maybe<Scalars['Boolean']['output']>;
  /** If the authenticated user if following this user */
  isFollowing: Maybe<Scalars['Boolean']['output']>;
  /** The user's media list options */
  mediaListOptions: Maybe<MediaListOptions>;
  /** The user's moderator roles if they are a site moderator */
  moderatorRoles: Maybe<Array<Maybe<ModRole>>>;
  /**
   * If the user is a moderator or data moderator
   * @deprecated Deprecated. Replaced with moderatorRoles field.
   */
  moderatorStatus: Maybe<Scalars['String']['output']>;
  /** The name of the user */
  name: Scalars['String']['output'];
  /** The user's general options */
  options: Maybe<UserOptions>;
  /** The user's previously used names. */
  previousNames: Maybe<Array<Maybe<UserPreviousName>>>;
  /** The url for the user page on the AniList website */
  siteUrl: Maybe<Scalars['String']['output']>;
  /** The users anime & manga list statistics */
  statistics: Maybe<UserStatisticTypes>;
  /**
   * The user's statistics
   * @deprecated Deprecated. Replaced with statistics field.
   */
  stats: Maybe<UserStats>;
  /** The number of unread notifications the user has */
  unreadNotificationCount: Maybe<Scalars['Int']['output']>;
  /** When the user's data was last updated */
  updatedAt: Maybe<Scalars['Int']['output']>;
};
// ...
export type EntyFragmentFragment = { id: number, status: MediaListStatus | null, progress: number | null, media: { id: number, episodes: number | null, status: MediaStatus | null, title: { english: string | null, romaji: string | null } | null, nextAiringEpisode: { airingAt: number, episode: number, timeUntilAiring: number } | null } | null };

export type MediaFragmentFragment = { id: number, episodes: number | null, status: MediaStatus | null, title: { english: string | null, romaji: string | null } | null, nextAiringEpisode: { airingAt: number, episode: number, timeUntilAiring: number } | null };

export type UserAnimeListQueryVariables = Exact<{
  userId: Scalars['Int']['input'];
}>;


export type UserAnimeListQuery = { MediaListCollection: { lists: Array<{ entries: Array<{ id: number, status: MediaListStatus | null, progress: number | null, media: { id: number, episodes: number | null, status: MediaStatus | null, title: { english: string | null, romaji: string | null } | null, nextAiringEpisode: { airingAt: number, episode: number, timeUntilAiring: number } | null } | null } | null> | null } | null> | null } | null };